Vivid Vibes A social Media Platform.
Description: 
Vivid Vibes is a Social Media Platform with A very Sleek Front End and an organized and sophisticated Back End.
The main idea is a Project designed to demonstrate the connection between the Front End of a web application and The Database Behind it.

Tools Used:
- Xampp Control Panel.
- Visual Studio Code

Languages:
- The code doesn't feature any framework.
- HTML
- CSS
- JavaScript
- PHP
- SQL

Overview:
Vivid Vibes features a social media platform that has a beautiful Front End and an accurate back End. The connectivity was established using PHP. Everything is being fetched in real time other than the images. For Images, I managed to code a custom algo, which enables different images to load for different users. not a complex logic but managed to accomplish. no doubt if real-time images are fetched we can call it done for now for this specific Project. 

Expected Improvements:
- Password Recovery
- Hashed Passwords for users security
- Realtime Image Fetching
- using advanced frameworks such as Laravel
- perhaps changing from PHP to Python since it seems outdated.